sonar refactoring finished
