--- src/app/reusable/avatar/avatar.component.html
@@ -1,6 +1,0 @@
-  <div class="image-container">
-    <img class="image" *ngIf="avatarUrl" [src]="avatarUrl" />
-    <div class="label h30" *ngIf="!avatarUrl && name" > {{name!  | nvInitials}} </div>
-  </div>
-
-
