+++ src/app/reusable/nv-button/nv-button.component.scss
@@ -9,20 +9,6 @@
   height: 40px;
   cursor: pointer;
 
-  &.type-Raised {
-    box-shadow: 0px 1px 1px;
-    .button {
-      .label {
-      }
-    }
-  }
-
-  &.type-Stroked {
-    .button {
-      .label {
-      }
-    }
-  }
 
   &.theme-color-Basic {
     background-color: $color-light-gray-22;
@@ -58,11 +44,5 @@
     }
   }
 
-  &.theme-color-Secondary {
-    .button {
-      .label {
-      }
-    }
-  }
 
 }
