+++ src/app/reusable/nv-checklist-dropdown/nv-checklist-dropdown.component.scss
@@ -176,9 +176,6 @@ $height: 40px;
   display: none !important;
 }
 
-// menu gap
-::ng-deep .mat-form-field-infix {}
-
 ::ng-deep .cdk-overlay-pane {
   margin-top: 12px;
 }
@@ -197,7 +194,6 @@ $height: 40px;
 }
 
 // disable material ico
-
 ::ng-deep .mat-button-wrapper {
   color: $color-light-gray-50 !important;
   margin-top: 5px !important ;
@@ -293,9 +289,6 @@ $height: 40px;
   height: 30px !important;
   line-height: 30px !important;
   margin-top: -8px !important;
-}
-::ng-deep  .mat-icon {
-  // margin-top: -8px;
 }
 
   ::ng-deep button {
