+++ src/app/reusable/nv-custom-datatable/nv-custom-datatable.component.html
@@ -1,7 +1,7 @@
 <ng-container>
 <div class="search-container">
-  <app-search class="search" [isAdvance]="isAdvance" [hideMenu]="hideMenu" [list]="tableTestData"
-  [searchableRefList]="searchableRefList" (filteredItems)="onItemsFilter($event)"></app-search> 
+  <app-nv-search class="search" [isAdvance]="isAdvance" [hideMenu]="hideMenu" [list]="tableTestData"
+  [searchableRefList]="searchableRefList" (filteredItems)="onItemsFilter($event)"></app-nv-search> 
 
 </div>
