+++ src/app/reusable/nv-custom-table/nv-custom-table.component.html
@@ -8,8 +8,8 @@
         <mat-icon matPrefix >search</mat-icon>
       </div>-->
       <div class="search-container">
-        <app-search class="search" [isAdvance]="isAdvance" [hideMenu]="hideMenu" [list]="tableTestData"
-        [searchableRefList]="searchableRefList" (filteredItems)="onItemsFilter($event)"></app-search> 
+        <app-nv-search class="search" [isAdvance]="isAdvance" [hideMenu]="hideMenu" [list]="tableTestData"
+        [searchableRefList]="searchableRefList" (filteredItems)="onItemsFilter($event)"></app-nv-search> 
       
       </div>
       <div class="table-scroll">
