+++ src/app/reusable/nv-date-picker/nv-date-picker.component.html
@@ -8,7 +8,7 @@
    </ng-container>
    <mat-datepicker-toggle matSuffix [for]="isDateRangePicker ? rangePicker : singlePicker">
      <mat-icon matDatepickerToggleIcon>
-       <img class="date-toggle-icon" src="./../../../assets/icons/date-picker.icon.svg">
+       <img class="date-toggle-icon" src="./../../../assets/icons/date-picker.icon.svg" alt='Toggle Icon'>
      </mat-icon>
    </mat-datepicker-toggle>
  </div>
