+++ src/app/reusable/nv-dropdown/nv-dropdown.component.html
@@ -4,7 +4,8 @@
       {{selectedItem | nvTrim: textTrimNumber}}
       <ng-container [ngTemplateOutlet]="arrowTemplate" [ngTemplateOutletContext]="{isArrowDown:isArrowDown}"></ng-container>
     </button>
-    <img *ngIf="fieldType === dropDownFieldType.Icon" class="dropdown-icon" data-toggle="dropdown"  src="./../../../assets/icons/plus-button.icon.svg">
+    <img *ngIf="fieldType === dropDownFieldType.Icon" class="dropdown-icon" data-toggle="dropdown" 
+     src="./../../../assets/icons/plus-button.icon.svg" alt='Plus'>
     <button *ngIf="fieldType === dropDownFieldType.Button" type="button" class="btn dropdown-button body20 label" data-toggle="dropdown" 
       (click)="isArrowDown=!isArrowDown" (blur)="onBlur()">
       {{'Dropdown Button'}}
@@ -18,8 +19,8 @@
 </div>
 
 <ng-template #arrowTemplate let-item="isArrowDown">
-  <img *ngIf="isArrowDown" class="arrow-icon"  src="{{arrowIcons?.downward}}">
-  <img *ngIf="!isArrowDown" class="arrow-icon" src="{{arrowIcons?.upward}}">
+  <img *ngIf="isArrowDown" class="arrow-icon"  src="{{arrowIcons?.downward}}" alt='Arrow'>
+  <img *ngIf="!isArrowDown" class="arrow-icon" src="{{arrowIcons?.upward}}" alt='Arrow'>
 </ng-template>
