+++ src/app/reusable/nv-multiselect-dropdown/nv-multiselect-dropdown.component.html
@@ -28,8 +28,8 @@
 
 <ng-template #arrowTemplate let-item="isArrowDown">
   <div class="arrow-icon-container" [ngClass]="{'no-item': !hasItem}">
-  <img *ngIf="isArrowDown" class="arrow-icon"  src="{{arrowIcons?.downward}}">
-  <img *ngIf="!isArrowDown" class="arrow-icon" src="{{arrowIcons?.upward}}">
+  <img *ngIf="isArrowDown" class="arrow-icon"  src="{{arrowIcons?.downward}}" alt='Arrow'>
+  <img *ngIf="!isArrowDown" class="arrow-icon" src="{{arrowIcons?.upward}}" alt='Arrow'>
 </div>
 </ng-template>
