--- src/app/reusable/nv-search/search.component.scss
@@ -1,219 +1,0 @@
-@import './../theme/base.scss';
-
-$height: 40px;
-$border-thickness: 1px;
-$icon-width: 16px;
-$padding: 12px;
-$icons-distance: 8px;
-$icons-distance-large: 18px;
-
-.dropdown {
-  position: relative;
-  padding: 0px;
-  margin: 0px;
-  width: 100%;
-  min-width: 100%;
-  height: $height;
-  border: 1px solid $color-white-10;
-  border-radius: 4px;
-
-  .form-outline {
-    position: relative;
-    height: 100%;
-    width: 100%;
-    display: flex;
-    flex-direction: row;
-    background: $color-white-20;
-    border-radius: 4px;
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-
-    &.label {
-      font-size: $font-size-x-small;
-      font-weight: $font-weight-normal;
-      line-height: $line-height-medium;
-      color: $color-dark-gray-12;
-    }
-    .icon-search {
-        position: relative;
-        display: flex;
-        align-items: center;
-        align-self: center;
-        justify-content: center;
-        margin-top: auto;
-        margin-bottom: auto;
-        cursor: pointer;
-        width: $icon-width;
-        height: $icon-width;
-        fill: $color-dark-gray-20;
-        color: $color-dark-gray-20;
-        margin-left: 12px;
-    }
-
-    .input-keyword-label {
-      position: relative;
-      min-width: 4px;
-      height: 100%;
-      width: min-content;
-      padding-left: $icons-distance;
-      background-color: transparent;
-      vertical-align: middle;
-      display: flex;
-      align-items: center;
-
-    }
-
-    .form-control {
-      position: relative;
-      height: 100%;
-      width: 100%;
-      background-color: transparent;
-      transition: none;
-      border: 1px solid $color-white-10;
-      border: none;
-      border-radius: 4px;
-      resize: none;
-      overflow: hidden;
-      padding-top: 7.5px;
-      padding-left: $icons-distance;
-
-      font-size: $font-size-x-small !important;
-      font-weight: $font-weight-normal !important;
-      line-height: $line-height-medium !important;
-      color: $color-dark-gray-12 !important;
-
-      &:focus {
-        outline: none;
-        border: 1px solid $color-light-gray-30;
-        box-shadow: none;
-        border-radius: 4px;
-        box-sizing: border-box;
-        padding-top: 7.5px;
-        padding-left:  $icons-distance;
-        border: none;
-      }
-
-      &::-webkit-input-placeholder {
-        font-family: $nv-font;
-        font-size: $font-size-x-small;
-        font-weight: $font-weight-normal;
-        line-height: $line-height-x-small;
-        color: $color-dark-gray-30;
-        padding-top: 4.5px;
-        padding-left: 0px;
-        opacity: 1;
-      }
-
-      &:-moz-placeholder {
-        font-family: $nv-font;
-        font-size: $font-size-x-small;
-        font-weight: $font-weight-normal;
-        line-height: $line-height-x-small;
-        color: $color-dark-gray-30;
-        padding-top: 4.5px;
-        padding-left: 0px;
-        opacity: 1;
-      }
-
-      &::-moz-placeholder {
-        font-family: $nv-font;
-        font-size: $font-size-x-small;
-        font-weight: $font-weight-normal;
-        line-height: $line-height-x-small;
-        color: $color-dark-gray-30;
-        padding-top: 4.5px;
-        padding-left: 0px;
-        opacity: 1;
-      }
-
-    }
-
-    .icon {
-      position: absolute;
-      top: 0px;
-      bottom: 0px;
-      display: flex;
-      align-items: center;
-      align-self: center;
-      justify-content: center;
-      margin-top: auto;
-      margin-bottom: auto;
-      cursor: pointer;
-      width: $icon-width;
-      height: $icon-width;
-      fill: $color-dark-gray-20;
-      color: $color-dark-gray-20;
-      opacity: 1;
-
-      &.search {
-        left: $padding;
-      }
-      &.cancel {
-        right: $icon-width + $padding + $icons-distance-large;
-      }
-      &.setting {
-        right: $padding;
-        height: 12px;
-      }
-    }
-
-  }
-
-  .dropdown-menu {
-    position: relative;
-    width: fit-content;
-    max-height: 320px;
-    min-width:  100%;
-    padding-top: 0px;
-    padding-bottom: 0px;
-    background: $color-white-20;
-    box-shadow: 0px 1px 4px $color-black-14;
-    border: 1px solid $color-white;
-    border-radius: 4px;
-
-    &.display-toggle-menu {
-      display: none;
-    }
-    &.display-menu {
-      display: none;
-    }
-
-    .dropdown-item {
-      display: flex;
-      align-items: center;
-      padding-left: 12px;
-      padding-right: 12px;
-      height: $height;
-
-      &.label {
-        letter-spacing: 0px;
-        opacity: 1;
-        font-style: normal;
-        font-size: $font-size-x-small;
-        font-weight: $font-weight-normal;
-        line-height: $line-height-xx-small;
-        color: $color-black-20;
-      }
-
-      &.selected-item {
-        font-style: normal;
-        font-weight: $font-weight-normal;
-        font-size: $font-size-x-small;
-        font-family: $nv-font;
-        line-height: $line-height-xx-small;
-        color: $color-black-20;
-        background-color: $color-light-gray;
-      }
-
-      &:hover {
-        background: $color-light-gray;
-      }
-    }
-
-  }
-
-}
-
-
-
