--- src/app/reusable/upload/upload.component.html
@@ -1,10 +1,0 @@
-<div class="container" appNvFileUpload (fileDropped)="onFileDropped($event)">
-  <div class="d-flex flex-column  align-items-center">
-    <input type="file" #fileDropRef id="fileDropRef" multiple (change)="fileBrowseHandler($event)" />
-    <img class="align-self-center mt-5" src="../../../assets/icons/ico.file.svg">
-    <div class="content-label h30 align-self-center mt-3 mb-3">Drag and drop file here</div>
-    <label for="fileDropRef h50 d-flex justify-content-center"> 
-      <span class="d-flex align-self-center">Upload</span></label>
-
-  </div>
-</div>
