--- src/app/reusable/upload/upload.component.scss
@@ -1,37 +1,0 @@
-@import './../theme/base.scss';
-
-.container {
-  position: relative;
-  width: 100%;
-  height: 100%;
-  text-align: center;
-  border: 1px dashed $color-dark-gray-10;
-  margin: 0 auto;
-
-  input {
-    position: absolute;
-    width: 100%;
-    height: 100%;
-    top: 0;
-    left: 0;
-    z-index: 2;
-    opacity: 0;
-  }
-.content-label {
-    color: $color-black-20;
-}
-  label {
-    position: relative;
-    display: flex;
-    justify-content: center;
-    width: 94px;
-    height: 40px;
-    background: $color-dark-blue-10 0% 0% no-repeat padding-box;
-    border: 1px solid $color-dark-blue;
-    color: $color-white-20;
-    border-radius: 4px;
-    opacity: 1;
-  }
-}
-
-
